pyproject.toml
[tool.poetry]
name = "hw14"
version = "0.1.0"
description = ""
authors = ["Tetiana <bovanenko17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.6"
fastapi = "^0.114.2"
psycopg2 = "^2.9.9"
libgravatar = "^1.0.4"
python-jose = "^3.3.0"
sqlalchemy = "^2.0.34"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
fastapi-mail = "^1.4.1"
redis = "^5.0.8"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.41.0"
aioredis = "^2.0.1"
python-multipart = "^0.0.9"
sphinx = "^8.0.2"
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
